ci(clippy): clear the lint backlog and make Clippy gate CI (#18)
The Clippy job was non-blocking (RUSTFLAGS="") to tolerate a ~460-lint
backlog (PR #10). This clears the backlog and re-enables it as a gate.

How the lints were resolved:
- ~188 mechanical lints auto-fixed via `cargo clippy --fix` (e.g.
  manual_is_multiple_of, int_plus_one, byte_char_slices, redundant
  closures, manual_ignore_case_cmp -> eq_ignore_ascii_case).
- 3 lints that are pedantic in this numerical/scientific codebase are
  allowed workspace-wide via [workspace.lints.clippy] in the root
  Cargo.toml (members opt in with [lints] workspace = true):
  needless_range_loop (202), too_many_arguments (17), type_complexity (7).
- The remaining ~50 genuine lints were hand-fixed (unnecessary_sort_by,
  manual_strip, vec_init_then_push, ptr_arg, manual_clamp, question_mark,
  while_let_loop, if_same_then_else, manual_div_ceil, nonminimal_bool,
  etc.), each behaviour-preserving. The ACMG combining-rule predicates in
  cyanea-omics keep an explicit form (targeted #[allow(nonminimal_bool)])
  so they mirror the published criteria.
- Fixed a pre-existing misspelled attribute (clippy::result_large_error_types
  -> clippy::result_large_err).

Note: CI's Clippy job resolves to the stable toolchain (dtolnay sets
RUSTUP_TOOLCHAIN, overriding rust-toolchain.toml), so a few lints only
stable/newer clippy emits — and x86_64-only SIMD lints in cyanea-align —
were included.

CI: the Clippy job now runs `cargo clippy --workspace -- -D warnings`
(RUSTFLAGS override removed). Verified: clippy clean (stable), fmt clean,
`cargo test --workspace --exclude cyanea-py` = 3653 passed, 0 failed.

+++ cyanea-epi/src/pileup.rs
@@ -85,11 +85,11 @@ pub fn build_pileup(reads: &[(String, u64, u64)], fragment_size: u64) -> TagPile
 
         // Extend coverage array if needed
         if end as usize >= cov.len() {
-            cov.resize((end as usize) + 1, 0);
+            cov.resize((end as usize) + 1, 0u32);
         }
 
         for i in *start as usize..=((end - 1) as usize).min(cov.len() - 1) {
-            cov[i] = (cov[i] as u32).saturating_add(1);
+            cov[i] = cov[i].saturating_add(1);
         }
     }
 
@@ -101,7 +101,7 @@ pub fn build_pileup(reads: &[(String, u64, u64)], fragment_size: u64) -> TagPile
 /// Normalization methods:
 /// - `"cpm"` — counts per million mapped reads
 /// - `"rpkm"` — reads per kilobase per million mapped reads
-#[allow(clippy::result_large_error_types)]
+#[allow(clippy::result_large_err)]
 pub fn normalize_pileup(pileup: &TagPileup, method: &str) -> Result<TagPileup> {
     let total_reads: u64 = pileup
         .coverage
